Cripsy Graphics, inlined Math, tasty subpixels!
This patch replaces the Matrix concatenation code
in Graphics.__update to a simplified version that
doesn’t lose precision due to less floating point
operations.

It `Math.floor`s the tx and ty positions, to have
pixel perfect position in the renderers. Any small
(subpixel) position difference is offset by the
__renderTransform, effectively doing high quality
smoothing in the software renderer.

The __width and __height are used by the Canvas
and Cairo renderer to determine buffer sizes.
These values are sized to fit the possible extra
pixel required for the subpixel rounding stuffs.

Before this patch:

- both the DOM and HTML5 WebGL renderer could show
  slight gaps and artifacts when rendering with
  roundPixels enabled (default).

_ `__worldTransform.a` and `d` would often end up
  being almost but not quite 1.000, resulting in
  blurry graphics by the GPU smoothing needlessly
  kicking in.